+++ up_tck_java/java_test_agent/src/main/java/org/tck/testagent/Main.java
@@ -1,8 +1,11 @@
-package org.tck;
+package org.tck.testagent;
 
 import com.google.protobuf.Any;
+import io.cloudevents.v1.proto.CloudEvent;
 import org.eclipse.uprotocol.v1.*;
 import org.json.JSONObject;
+import org.tck.up_client_socket_java.SocketUListener;
+import org.tck.up_client_socket_java.SocketUTransport;
 
 import java.io.IOException;
 import java.net.Socket;
