+++ up_tck_java/java_test_agent/src/main/java/org/tck/up_client_socket_java/SocketUTransport.java
@@ -1,4 +1,4 @@
-package org.tck;
+package org.tck.up_client_socket_java;
 
 import java.io.IOException;
 import java.io.InputStream;
